Writes data to JSON file, saves fragments in QA directory
This set of changes writes all the fragment information to a JSON file
(samples.json) which is saved in the Work directory as well as in the
data directory of the QA tool. This makes it easy to tweak fragments in
the future without having to manually change the QA tool.

Fragments are now saved in the Work directory as well as in the data
directory of the QA tool (so I never forget to copy across changes).

NB There is still a bug in the code which means it doesn't handle
multiple streets - so safest for now to just read in one street at a time
which isn't a problem given the way I am incrementally creating these
unique streets with the items I want on them.