Support for enchanted tree, gardening vendor and zutto spirit
Added in code for these 3 items so can generate the fragments. Only
remaining item are garden plots - but they come in different
sizes/styles before you even take account of the plants, so probably not
much use doing a compare for an item which is very unlikely to match my
sample.