Default offset/image sizes using existing files if found
When generating new images from a better street snap it is annoying to
have to manually enter the offset/size information. So now this is used
as an initial starting point, allowing the user to just press 's' to
save the image.

Item images are now stored in a user directory rather than in the data
directory.