Generating images of items which change according to maturity
Trees, jellisacs and barnacles all have different images depending on
the maturity. This program will now generate images of the form
tsid_state or tsid_variant_state so that snaps can be searched with all
images of all trees so that immature trees can be located and give a
useful x,y.

This has meant that transparency is now needed on fragment images so
that an entire tree trunk can be used for the reference without
background getting in the way. When comparing the fragment against all
other tree images, if the fragment has a transparent pixel,  then the
comparison is skipped for that pixel.

A line is now drawn on the reference picture to show where pigs come up
to, so ideally, fragments are chosen to lie above that line. So that
pigs don't interfere with matching.

Width/height of fragment is now recorded even though it is not used by
QABot - useful information when I am reconstructing fragments.