Handle duplicate matches for wood trees
Many of the trees produce duplicate matches for more mature versions of
the same variant. E.g. variant 1 maturity 3 will match maturity 4
because the only difference is a background branch in maturity 4. As the
maturity 3 fragment has transparent pixels where the background branch
is, then these pixels are skipped during the comparison of the 2
fragments. So they produce identical matches. This has only come to
light now that I have automatically removed all pixels affected by the
background colour - thereby producing more perfect complete image
samples to work from. So a match is ignored if found for an older
version of the same variant.