Mode to record min/max heights of quoins, support for enchanted wood trees
With the quoin_heights_only option, the tool will log the max/min
y_offset for quoins to a new JSON file - no fragment will be saved.
Then, when populating the samples.json file, the y_offset will be set to
midway between the two. It is anticipated that this option will only be
used once - and from then onwards the values read if quoin fragments are
regenerated. The value of y_offset is only recorded if the fragment
selected exactly matches the existing fragment previously generated with
this tool.

This means that the UniqueFragmentCheck class now does something
slightly different in this case - compares the selected fragment against
the existing quoin fragment (rather than street fragment).

This tool now also handles enchanted wood trees.