Added ? and S options
Sometimes useful to see the results of a possible save without it being
done - use ?. The S option allows the current image to be saved even if
more than one image has a match - to deal with complexity of wood trees
which can have identical fragments in two trees of the same variant. As
this is allowable, useful to be able to save when multiple perfect
matches, so long as the fragment includes an additional branch in the
next mature level.