Removed failed attempt at zutto vendor
As the 4 different vendors all fail to give a good match on a real
street, I took a sample instead from the base. As it was a perfect match
for all 4 images, I created a street_spirit_zutto.png. But this also
failed the real test with QABot as zoi's snap revealed a different base
pattern to my version. So QABot now skips these vendors. But left the
code for the different variants in this tool.